[#36] feat: confirmModal 추가

+++ src/hooks/useModal.tsx
@@ -11,6 +11,8 @@ export default function useModal() {
       title: <h3 className="f20 font-bold text-text_primary">{title}</h3>,
       content: <p className="f14 font-medium text-text_secondary">{content}</p>,
       icon: null,
+      maskClosable: true,
+      okText: '확인',
       okButtonProps: {
         style: {
           backgroundColor: '#626FE5',
@@ -20,10 +22,40 @@ export default function useModal() {
         },
       },
       onOk: () => {
-        Modal.destroyAll();
         onClick();
+        Modal.destroyAll();
       },
     });
   };
-  return { modal };
+  const confirmModal = ({ title, content, onClick = () => null }: ModalProps) => {
+    Modal.confirm({
+      title: <h3 className="f20 font-bold text-text_primary">{title}</h3>,
+      content: <p className="f14 font-medium text-text_secondary">{content}</p>,
+      icon: null,
+      maskClosable: true,
+      okText: '확인',
+      cancelText: '취소',
+      okButtonProps: {
+        style: {
+          backgroundColor: '#626FE5',
+          color: 'white',
+          borderRadius: '4px',
+          border: 'none',
+        },
+      },
+      cancelButtonProps: {
+        style: {
+          backgroundColor: '#F5F6FA',
+          color: '#626FE5',
+          borderRadius: '4px',
+          border: 'none',
+        },
+      },
+      onOk: () => {
+        onClick();
+        Modal.destroyAll();
+      },
+    });
+  };
+  return { modal, confirmModal };
 }
